Camera calibration process
Please check the word document describing all the process and the answer
on the question number 3. You can add things for sure if you want, or
repeat the process. I guess the big error comes from the difficulty of
clicking on the correct corner edges, and from the large images
resolution.

If you want, you can tranfer it to the latex format.
If not, we can simply export it to pdf, and then glue several pdf
documents together.